Move `calc_t_out` to `utils` + some cleanup
